printf

+++ _octal.c
@@ -7,7 +7,7 @@
 **/
 int _octal(va_list args)
 {
-	unsigned long int n = va_arg(args, unsigned int);
+	unsigned long int n = va_arg(args, unsigned long int);
 	char arr[1024];
 	int i = 0;
 	int j;
